Move PackedPrice & UnpackedDoubleU224
